待测试修正，主问题在DefaultSqlSession的parameter上

+++ behavior/src/main/java/com/github/dolphinmind/behavior/mediator/dao/IAward.java
@@ -11,6 +11,6 @@ import com.github.dolphinmind.behavior.mediator.po.Award;
 
 public interface IAward {
 
-    Award queryAwardInfoByAwardId(Long awardId);
+    Award queryAwardInfoByAwardId(Integer awardId);
 
 }
